add support of kirin710 model (huawei)
Allows all phone models that have a kirin 710 processor to be seen as huawei branded phones.

The following persist.sys.overlay.huawei property will be set to yes if the ro.hardware property or the ro.vendor.build.fingerprint property contains kirin710.

persist.sys.overlay.huawei is used in the rw-system.sh file (for the NFC conf), in the huawei overlay and also at the trebleapp level